Corrected implementation of class, password locked VI are now class members, they are implemented in correct way, original ones were moved into VIs_NoClassImplementation/ folder to have them in library in any case :)
There was problem that transcriptor defines basic transcriptor object from which all libraries should inherit if they want to be used in conjuction with transcriptor. But this basic object is defined in transcriptor, so we have to somehow include it into library. For this reason I put transcriptor basic object control and indicator on pallete, so basically user has to install transcriptor package before implemnting and extending libraries.
I added VIs for all members functions of Servo library, but need to define them in Translator.vi